apps/backend/requirements.txt
# Core - Updated for Python 3.13 compatibility
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database - MongoDB - Compatible versions
motor==3.6.0
pymongo==4.9.1
beanie==1.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Validation - Using newer versions with Python 3.13 support
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utilities
python-dateutil==2.8.2
cloudinary==1.41.0

# Excel/CSV Import
openpyxl==3.1.5

# Google Sheets Sync
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0

# Development dependencies - Updated versions
black==24.10.0
isort==5.13.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
